More progress :>

+++ Logic/source/AI/Behavior/BigBadBehavior.cpp
@@ -32,7 +32,7 @@ BigBadBehavior::BigBadBehavior() :
     });
 
     node = addNode(parent, NodeType::CONDITION, 2, [](RunIn &in) -> bool {
-        return (in.enemy->getHealth() / in.enemy->getMaxHealth()) > 0.30f;
+        return (static_cast<float> (in.enemy->getHealth()) / in.enemy->getMaxHealth()) > 0.30f;
     });
     addNode(node, NodeType::ACTION, 0, [](RunIn &in) -> bool {
         in.enemy->useAbility(*in.target, PHASE_TWO);
